+++ Changes.md
@@ -1,4 +1,15 @@
-﻿## version BIT.Data.Sync 1.0.1
+﻿## version 7.0.3.2
+    -Added extension AddSyncFrameworkForMysql
+    -Added extension AddSyncFrameworkForNpgsql
+    -Added extension AddSyncFrameworkForSqlite
+    -Added extension AddSyncFrameworkForSqlServer
+
+    -Added BIT.Data.Sync.AspNetCore to host the controlers and extension for the sync server
+
+    -Added AddSyncServerWithMemoryNode to add a memory node to the sync server
+
+## version 7.0.3.1
+## version BIT.Data.Sync 1.0.1
 ### ISequenceService
     - Added new interface ISequenceService that is on charge or generating sequence numbers for delta entities
     - Added new EfSquenceService  (ISequenceService implementation)